feat(admin): 2253 - pdt cohort switcher
