+++ admin/src/scenes/plan-transport/table-repartition/index.jsx
@@ -1,14 +1,14 @@
 import React from "react";
 import { Switch } from "react-router-dom";
 import { SentryRoute } from "../../../sentry";
-import National from "./National";
-import Regional from "./Regional";
+import NationalView from "./NationalView";
+import RegionalView from "./RegionalView";
 
 export default function TableRepartition() {
   return (
     <Switch>
-      <SentryRoute path="/table-repartition/regional" component={Regional} />
-      <SentryRoute path="/table-repartition" component={National} />
+      <SentryRoute path="/table-repartition/regional" component={RegionalView} />
+      <SentryRoute path="/table-repartition" component={NationalView} />
     </Switch>
   );
 }
